Porting Jmetal5 to OPLA-tool
Ported PLACrossover, PLAMutation and OPLAProblem to jmetal5

implemented a try-run class NSGAII_Jmetal5

+++ opla-core/src/main/java/jmetal/metaheuristics/nsgaII/NSGAII.java
@@ -72,8 +72,8 @@ public class NSGAII extends Algorithm {
         Distance distance = new Distance();
 
         // Read the parameters
-        populationSize = ((Integer) getInputParameter("populationSize")).intValue();
-        maxEvaluations = ((Integer) getInputParameter("maxEvaluations")).intValue();
+        populationSize = (Integer) getInputParameter("populationSize");
+        maxEvaluations = (Integer) getInputParameter("maxEvaluations");
         indicators = (QualityIndicator) getInputParameter("indicators");
 
         // Initialize the variables
